feat: Scaffold DarcyAi Android Application
- Initialize Android project structure for `DarcyAi` under a new root directory.
- Configure Gradle with `compose-markdown` (JitPack) and Wrapper (8.0).
- Implement "Neural Neon" UI (Compose) with custom components.
- Implement `HiveManager` for NPU-aware model swapping.
- Add `ControlRoomService` for network syncing.
- Fix Gradle Kotlin DSL ambiguity for `platform()` calls by adding explicit casts.
- Update `MainActivity` to use correct Material3 `TextFieldDefaults.colors` parameters.
- Clean up `externalNativeBuild` to prevent configuration errors.
- Ensure type safety in `NeonComponents.kt` calculations.
- Add `.gitignore`.

+++ DarcyAi/app/src/main/java/com/darcyai/ui/components/NeonComponents.kt
@@ -73,9 +73,9 @@ fun StarTrigger(
             for (i in 0 until numPoints * 2) {
                 val radius = if (i % 2 == 0) outerRadius else innerRadius
                 val angle = Math.PI / numPoints * i - Math.PI / 2
-                val x = cx + cos(angle) * radius
-                val y = cy + sin(angle) * radius
-                if (i == 0) path.moveTo(x.toFloat(), y.toFloat()) else path.lineTo(x.toFloat(), y.toFloat())
+                val x = cx + (cos(angle) * radius).toFloat()
+                val y = cy + (sin(angle) * radius).toFloat()
+                if (i == 0) path.moveTo(x, y) else path.lineTo(x, y)
             }
             path.close()
